TimeSlot policy

+++ app/controllers/time_slots_controller.rb
@@ -1,6 +1,8 @@
 class TimeSlotsController < ApplicationController
   def show
     @school     = School.find(params[:school_id])
+    authorize @school, :update?
+    
     @date       = (params[:date] && Date.parse(params[:date])) || Date.today
     @time_slots = TimeSlot.for_school(@school).for_day(@date)
 
@@ -11,8 +13,10 @@ class TimeSlotsController < ApplicationController
   end
 
   def create
-    @school     = School.find(params[:school_id])
-    @date       = (params[:date] && Date.parse(params[:date])) || Date.today
+    @school = School.find(params[:school_id])
+    authorize @school, :update?
+
+    @date = (params[:date] && Date.parse(params[:date])) || Date.today
     school_year = SchoolYear.find(time_slot_params[:school_year_id])
 
     @slots = TimeSlot.create_week_daily(school_year: school_year, date: @date, params: time_slot_params)
@@ -25,6 +29,9 @@ class TimeSlotsController < ApplicationController
 
   def update
     @slot = TimeSlot.find(params[:id])
+    @school = @slot.school
+    authorize @school, :update?
+
     TimeSlot.update_with_type(time_slot: @slot, type: params[:type], params: update_slot_params)
     @date = @slot.start.to_date
 
@@ -36,6 +43,8 @@ class TimeSlotsController < ApplicationController
   def destroy
     @slot = TimeSlot.find(params[:id])
     @school = @slot.school
+    authorize @school, :update?
+
     @date = @slot.start.to_date
     TimeSlot.destroy_with_type(time_slot: @slot, type: params[:type])
